api_onson_mail/requirements.txt
Django==4.1.5
psycopg2-binary==2.9.5
python-dotenv==0.21.0
redis
pillow
django-import-export==3.3.9
tablib[xlsx]
phonenumbers==8.12.54
django-parler
django-jet-reboot==1.3.3
django-simple-history
django-jazzmin==2.6.0
djangorestframework==3.13.1
drf-spectacular==0.24.1
django-cors-headers==3.13.0
cryptography
pyjwt
requests
gunicorn
celery==5.2.7
django-celery-results
djangorestframework-simplejwt
django-filter
django-storages
boto3
django-phonenumber-field
xhtml2pdf
qrcode==7.4.2